--- a/scanner.go
+++ b/scanner.go
@@ -161,7 +161,6 @@ func (s *Scanner) parseString() int {
 func (s *Scanner) parseNumber() int {
 	const (
 		begin = iota
-		sign
 		leadingzero
 		anydigit1
 		decimal
@@ -175,18 +174,16 @@ func (s *Scanner) parseNumber() int {
 	w := s.br.window(0)
 	// int vs uint8 costs 10% on canada.json
 	var state uint8 = begin
+	if w[0] == '-' {
+		pos++
+		w = s.br.window(1)
+	}
+
 done:
 	for {
 		for _, elem := range w {
 			switch state {
 			case begin:
-				// only accept sign or digit
-				if elem == '-' {
-					state = sign
-					break
-				}
-				fallthrough
-			case sign:
 				if elem >= '1' && elem <= '9' {
 					state = anydigit1
 				} else if elem == '0' {
@@ -202,14 +199,15 @@ done:
 				}
 				fallthrough
 			case leadingzero:
-				switch elem {
-				case '.':
+				if elem == '.' {
 					state = decimal
-				case 'e', 'E':
-					state = exponent
-				default:
-					break done
+					break
 				}
+				if elem == 'e' || elem == 'E' {
+					state = exponent
+					break
+				}
+				break done
 			case decimal:
 				if elem >= '0' && elem <= '9' {
 					state = anydigit2
@@ -221,13 +219,12 @@ done:
 				if elem >= '0' && elem <= '9' {
 					break
 				}
-				switch elem {
-				case 'e', 'E':
+				if elem == 'e' || elem == 'E' {
 					state = exponent
-				default:
-					s.pos = pos
-					return pos // finished
+					break
 				}
+				s.pos = pos
+				return pos // finished.
 			case exponent:
 				if elem == '+' || elem == '-' {
 					state = expsign
@@ -237,10 +234,10 @@ done:
 			case expsign:
 				if elem >= '0' && elem <= '9' {
 					state = anydigit3
-				} else {
-					// error
-					return -1
+					break
 				}
+				// error
+				return -1
 			case anydigit3:
 				if elem < '0' || elem > '9' {
 					break done