--- a/scanner_test.go
+++ b/scanner_test.go
@@ -164,7 +164,7 @@ func BenchmarkParseNumber(b *testing.B) {
 						r:    r,
 					},
 				}
-				n := scanner.parseNumber()
+				n := scanner.parseNumber(scanner.br.window(0)[0])
 				if n != len(tc) {
 					b.Fatalf("failed")
 				}