```
benchstat {old,new}.txt
name                                     old time/op    new time/op    delta
DecoderNextToken/pkgjson/canada-4          7.25ms ± 0%    7.09ms ± 0%  -2.17%  (p=0.008 n=5+5)
DecoderNextToken/pkgjson/citm_catalog-4    3.11ms ± 1%    3.01ms ± 1%  -3.23%  (p=0.008 n=5+5)
DecoderNextToken/pkgjson/twitter-4         1.48ms ± 1%    1.48ms ± 1%    ~     (p=0.690 n=5+5)
DecoderNextToken/pkgjson/code-4            7.30ms ± 1%    7.26ms ± 1%    ~     (p=0.222 n=5+5)
DecoderNextToken/pkgjson/example-4         30.9µs ± 0%    31.3µs ± 1%  +1.22%  (p=0.008 n=5+5)
DecoderNextToken/pkgjson/sample-4           666µs ± 1%     660µs ± 0%  -0.98%  (p=0.008 n=5+5)

name                                     old speed      new speed      delta
DecoderNextToken/pkgjson/canada-4         311MB/s ± 0%   317MB/s ± 0%  +2.21%  (p=0.008 n=5+5)
DecoderNextToken/pkgjson/citm_catalog-4   556MB/s ± 1%   575MB/s ± 1%  +3.33%  (p=0.008 n=5+5)
DecoderNextToken/pkgjson/twitter-4        428MB/s ± 1%   427MB/s ± 1%    ~     (p=0.643 n=5+5)
DecoderNextToken/pkgjson/code-4           266MB/s ± 1%   267MB/s ± 1%    ~     (p=0.222 n=5+5)
DecoderNextToken/pkgjson/example-4        421MB/s ± 0%   416MB/s ± 1%  -1.20%  (p=0.008 n=5+5)
DecoderNextToken/pkgjson/sample-4        1.03GB/s ± 1%  1.04GB/s ± 0%  +0.99%  (p=0.008 n=5+5)
```